implemented controllers for users
